refactor(api): modularizar resto de endpoints y funciones de utilidad
